Keyboard focus indicator is toggled on focus method
The focus rectangle is only displayed when the control was focused through a keyboard action, and it is hidden again when a mouse action has occured. This should be the default behaviour for Windows controls and be similar to what FocusVisualStyle would have done if it was available for this usage.