Fixed selection border when scrolled down; exact border selection
When the TreeViewEx was scrolled down, the selection border was painted at the absolute position inside the ScrollViewer, hence not where the mouse cursor was. Since the selection border is now outside the ScrollViewer, none of its previous padding and border width correction is needed anymore.

Also, the border selection now works exactly. (The right edge is not the left + the width, but one less.)