Renamed from TreeViewEx to MultiSelectTreeView
To better distinguish this fork from the original TreeViewEx project, and to allow for a less complicated publishing of my fork, I have decided that it would be best to rename the project altogether. This way the project name is unique again. I am going to continue to work on and with my fork because the differences between Goroll's TreeViewEx and my version have grown too large to merge and our opinions, Goroll's and mine, may be different. Occasional synchronisation of bug fixes or enhancements between both project may still occur in the future.